Disk: Moved disk files into VFS dir
