+++ lib/alloc/src/string/mod.rs
@@ -23,7 +23,7 @@ DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
 OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-use core::fmt::{Formatter, Write};
+use core::fmt::Write;
 use core::ops::Deref;
 use crate::heap::{AllocatorAPI, GlobalAlloc};
 use crate::vec::Vec;
